week_2/day_5/project/requirements.txt
Flask>=2.2
psycopg[binary]>=3.1
python-dotenv>=1.0
googletrans==4.0.0-rc1
gTTS>=2.3.0